Shruti_102218077/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
faiss-cpu==1.7.4
sentence-transformers==3.0.1
transformers==4.44.2
torch>=2.0.0
torchvision>=0.15.0
Pillow==10.4.0
requests==2.32.3
matplotlib==3.8.4
langchain==0.2.14
langchain-community==0.2.10
tiktoken==0.7.0
pinecone-client==5.0.1
qdrant-client==1.9.2
pyngrok==7.2.3
gdown==5.2.0
